Thing CRUD gets dual navigation

+++ www/tracking/routing/thing_routes.py
@@ -4,11 +4,12 @@ from flask import Blueprint, request, redirect
 from flask_login import login_required, current_user
 
 from tracking import database
-from tracking.navigation.cupboard_navigation import create_cupboard_navigator
 from tracking.forms.thing_forms import ThingUpdateForm, ThingCreateForm, update_thing_from_form
+from tracking.modelling.place_model import find_place_by_id
 from tracking.modelling.thing_model import find_thing_by_id
 from tracking.navigation.dual_navigator import DualNavigator
 from tracking.routing.home_redirect import home_redirect
+from tracking.viewing.cupboard_display_context import CupboardDisplayContext
 
 thing_bp = Blueprint(
     'thing_bp', __name__,
@@ -17,31 +18,32 @@ thing_bp = Blueprint(
 )
 
 
-@thing_bp.route('/create/<int:thing_id>', methods=['POST', 'GET'])
+@thing_bp.route('/create/<int:place_id>/<int:thing_id>', methods=['POST', 'GET'])
 @login_required
-def thing_create(thing_id):
+def thing_create(place_id, thing_id):
+    place = find_place_by_id(place_id)
     thing = find_thing_by_id(thing_id)
-    if thing is None or not thing.may_create_thing(current_user):
+    if place and thing is None or not thing.may_create_thing(current_user):
         return home_redirect()
     form = ThingCreateForm()
-    navigator = DualNavigator(thing.root)
+    navigator = DualNavigator(place=place, thing=thing)
     if request.method == 'POST' and form.cancel_button.data:
         return redirect(navigator.url(thing, 'view'))
     if form.validate_on_submit():
         new_thing = thing.create_kind_of_thing(name=form.name.data, description=form.description.data)
         return redirect(navigator.url(new_thing, 'view'))
     else:
-        return thing.display_context(navigator, current_user).render_template('pages/form_page.j2', form=form,
-                                                                              form_title=f'Create New Kind of '
-                                                                                         f'{thing.name}')
+        return CupboardDisplayContext().render_template(
+            'pages/form_page.j2', form=form, form_title=f'Create New Kind of {thing.name}')
 
 
-@thing_bp.route('/delete/<int:thing_id>')
+@thing_bp.route('/delete/<int:place_id>/<int:thing_id>')
 @login_required
-def thing_delete(thing_id):
+def thing_delete(place_id, thing_id):
+    place = find_place_by_id(place_id)
     thing = find_thing_by_id(thing_id)
-    if thing is not None and thing.may_delete(current_user):
-        navigator = create_cupboard_navigator()
+    if place and thing is not None and thing.may_delete(current_user):
+        navigator = DualNavigator(place=place, thing=thing)
         redirect_url = navigator.url(thing.parent_object, 'view')
         database.session.delete(thing)
         database.session.commit()
@@ -50,13 +52,14 @@ def thing_delete(thing_id):
         return home_redirect()
 
 
-@thing_bp.route('/update/<int:thing_id>', methods=['GET', 'POST'])
+@thing_bp.route('/update/<int:place_id>/<int:thing_id>', methods=['GET', 'POST'])
 @login_required
-def thing_update(thing_id):
+def thing_update(place_id, thing_id):
+    place = find_place_by_id(place_id)
     thing = find_thing_by_id(thing_id)
-    if thing and thing.may_update(current_user):
+    if place and thing and thing.may_update(current_user):
         form = ThingUpdateForm(obj=thing)
-        navigator = create_cupboard_navigator()
+        navigator = DualNavigator(place=place, thing=thing)
         redirect_url = navigator.url(thing, 'view')
         if request.method == 'POST' and form.cancel_button.data:
             return redirect(redirect_url)
@@ -65,8 +68,7 @@ def thing_update(thing_id):
             database.session.commit()
             return redirect(redirect_url)
         else:
-            return thing.display_context(navigator, current_user).render_template('pages/form_page.j2', form=form)
+            return CupboardDisplayContext().render_template(
+                'pages/form_page.j2', form=form, form_title=f'Update {thing.name}')
     else:
         return home_redirect()
-
-
